a step towards 'green' tests

+++ modules/core/src/array.cpp
@@ -267,7 +267,7 @@ cvInitMatNDHeader( CvMatND* mat, int dims, const int* sizes,
         mat->dim[1].size = dims == 0 ? 1 : mat->dim[0].size;
         mat->dim[1].step = (int)esz;
         mat->dim[0].size = 1;
-        mat->dim[0].step = mat->dim[1].size*esz;
+        mat->dim[0].step = (int)(mat->dim[1].size*esz);
     }
     mat->data.ptr = (uchar*)data;
     mat->refcount = 0;
